Use int32 for remote console buffered count
_sim_rem_log_out stores the result of tmxr_send_buffered_data and
compares it with TMLN.txbsz. Both of those values are int32 counts, so
use int32 for the local variable instead of size_t.

This matches the existing TMXR count domain and removes the
sign-comparison warning without changing the buffering logic.

+++ src/runtime/sim_console.c
@@ -903,7 +903,7 @@ if ((!sim_oline) && (sim_log)) {
 sim_oline = NULL;
 if ((rem->act == NULL) &&
     (!tmxr_input_pending_ln (lp))) {
-    size_t unwritten;
+    int32 unwritten;
 
     do {
         unwritten = tmxr_send_buffered_data (lp);
